requirements.txt
streamlit
azure-ai-translation-text
azure-core
python-dotenv
fastapi
uvicorn
python-multipart